pyproject.toml
[project]
name = "elementalBack"
version = "0.1.0"
description = "Elemental Backend"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.2",
    "aiosmtplib>=5.0.0",
    "aiosqlite>=0.21.0",
    "alembic>=1.15.2",
    "authlib>=1.5.2",
    "bcrypt==3.2.2",
    "fastapi[standard]>=0.115.12",
    "itsdangerous>=2.2.0",
    "passlib>=1.7.4",
    "pydantic-settings>=2.9.1",
    "pyjwt>=2.10.1",
    "python-magic>=0.4.27",
    "sqlalchemy>=2.0.40",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "aiosqlite>=0.21.0",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
    "bandit>=1.9.3",
    "wapiti3>=3.2.10",
]


[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
pythonpath = [
    "."
]
filterwarnings = [
    "ignore"
]